feat: add support for depends_on optoin(dependency resolution)
refactor: refactor folder structure and add more examples
