+++ grid-compose/.gitignore
@@ -2,3 +2,4 @@ bin/*
 .pre-commit-config.yaml
 out
 full_example.yml
+invalid
