+++ grid-compose/cmd/root.go
@@ -1,19 +1,13 @@
 package cmd
 
 import (
+	"context"
 	"os"
 
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
-	"github.com/threefoldtech/tfgrid-sdk-go/grid-compose/internal"
-)
-
-var (
-	app        *internal.App
-	configPath string
-	network    string
-	mnemonic   string
+	"github.com/threefoldtech/tfgrid-sdk-go/grid-compose/internal/app"
 )
 
 func Execute() {
@@ -22,23 +16,35 @@ func Execute() {
 	}
 }
 
-// TODO: Validate command line arguments
+// TODO: validate command line arguments
 var rootCmd = &cobra.Command{
 	Use:   "grid-compose",
 	Short: "Grid-Compose is a tool for running multi-vm applications on TFGrid defined using a Yaml formatted file.",
 	PersistentPreRun: func(cmd *cobra.Command, args []string) {
-		var err error
-		app, err = internal.NewApp(network, mnemonic, configPath)
+		network := os.Getenv("NETWORK")
+		mnemonic := os.Getenv("MNEMONIC")
+		configPath, _ := cmd.Flags().GetString("file")
+
+		app, err := app.NewApp(network, mnemonic, configPath)
+
 		if err != nil {
 			log.Fatal().Err(err).Send()
 		}
+
+		ctx := context.WithValue(cmd.Context(), "app", app)
+		cmd.SetContext(ctx)
 	},
 }
 
 func init() {
-	network = os.Getenv("NETWORK")
-	mnemonic = os.Getenv("MNEMONIC")
-	rootCmd.PersistentFlags().StringVarP(&configPath, "file", "f", "./grid-compose.yaml", "the grid-compose configuration file")
+	rootCmd.PersistentFlags().StringP("file", "f", "./grid-compose.yaml", "the grid-compose configuration file")
+
+	rootCmd.AddCommand(downCmd)
+	rootCmd.AddCommand(upCmd)
+	rootCmd.AddCommand(versionCmd)
+
+	psCmd.PersistentFlags().BoolP("verbose", "v", false, "all information about deployed services")
+	rootCmd.AddCommand(psCmd)
 
 	log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr})
 }
