+++ grid-compose/examples/dependency/multiple_dependencies.yml
@@ -1,16 +1,6 @@
 version: '1.0.0'
 
 networks:
-  net1:
-    name: 'miaminet'
-    range:
-      ip:
-        type: ipv4
-        ip: 10.20.0.0
-      mask:
-        type: cidr
-        mask: 16/32
-      wg: true
   net2:
     name: 'cartoonnetwork'
     range:
@@ -24,24 +14,17 @@ networks:
 services:
   server:
     flist: 'https://hub.grid.tf/tf-official-vms/ubuntu-24.04-full.flist'
+    ip_types:
+      - ipv4
+      - ygg
     entrypoint: '/sbin/zinit init'
     volumes:
       - webdata
     resources:
       cpu: 1
       memory: 2048
-      rootfs: 2048
+      rootfs: 25600
     network: net2
-  server2:
-    flist: 'https://hub.grid.tf/tf-official-vms/ubuntu-24.04-full.flist'
-    entrypoint: '/sbin/zinit init'
-    volumes:
-      - webdata
-    resources:
-      cpu: 1
-      memory: 2048
-      rootfs: 2048
-    network: net1
   frontend:
     flist: 'https://hub.grid.tf/tf-official-vms/ubuntu-24.04-full.flist'
     ip_types:
@@ -52,8 +35,12 @@ services:
     resources:
       cpu: 1
       memory: 2048
-      rootfs: 2048
-    network: net1
+      rootfs: 25600
+    network: net2
+    node_id: 144
+    depends_on:
+      - server
+      - database
   database:
     flist: 'https://hub.grid.tf/tf-official-vms/ubuntu-24.04-full.flist'
     ip_types:
@@ -64,7 +51,7 @@ services:
     resources:
       cpu: 1
       memory: 2048
-      rootfs: 2048
+      rootfs: 25600
     network: net2
 
 volumes:
