+++ grid-compose/examples/single_service_3.yml
@@ -19,6 +19,6 @@ services:
     resources:
       cpu: 1
       memory: 2048
-      rootfs: 25600
-    node_id: 14
+      rootfs: 2048
+    node_id: 144
     network: net1
