+++ grid-compose/examples/two_services_same_network_2.yml
@@ -9,11 +9,11 @@ services:
     entrypoint: '/sbin/zinit init'
     volumes:
       - webdata
-    node_id: 14
+    node_id: 144
     resources:
       cpu: 1
       memory: 2048
-      rootfs: 25600
+      rootfs: 2048
 
   database:
     flist: 'https://hub.grid.tf/tf-official-vms/ubuntu-24.04-full.flist'
@@ -25,7 +25,7 @@ services:
     resources:
       cpu: 1
       memory: 2048
-      rootfs: 25600
+      rootfs: 2048
 
 volumes:
   webdata:
