+++ grid-compose/internal/app/version.go
@@ -1,0 +1,1 @@
+package app
