+++ grid-compose/internal/types/service.go
@@ -12,8 +12,6 @@ type Service struct {
 	Network     string       `yaml:"network"`
 	HealthCheck *HealthCheck `yaml:"healthcheck,omitempty"`
 	DependsOn   []string     `yaml:"depends_on,omitempty"`
-
-	Name string
 }
 
 // Resources represents the resources required by the service
@@ -35,31 +33,4 @@ type HealthCheck struct {
 type Volume struct {
 	MountPoint string `yaml:"mountpoint"`
 	Size       string `yaml:"size"`
-}
-
-// Network represents the network configuration
-type Network struct {
-	Name         string            `yaml:"name"`
-	Description  string            `yaml:"description"`
-	IPRange      IPNet             `yaml:"range"`
-	AddWGAccess  bool              `yaml:"wg"`
-	MyceliumKeys map[uint32][]byte `yaml:"mycelium_keys"`
-}
-
-// IPNet represents the IP and mask of a network
-type IPNet struct {
-	IP   IP     `yaml:"ip"`
-	Mask IPMask `yaml:"mask"`
-}
-
-// IP represents the IP of a network
-type IP struct {
-	Type string `yaml:"type"`
-	IP   string `yaml:"ip"`
-}
-
-// IPMask represents the mask of a network
-type IPMask struct {
-	Type string `yaml:"type"`
-	Mask string `yaml:"mask"`
 }
