--- grid-compose/internal/utils/mycelium_seed.go
@@ -1,13 +1,0 @@
-package utils
-
-import (
-	"crypto/rand"
-
-	"github.com/threefoldtech/zos/pkg/gridtypes/zos"
-)
-
-func GetRandomMyceliumIPSeed() ([]byte, error) {
-	key := make([]byte, zos.MyceliumIPSeedLen)
-	_, err := rand.Read(key)
-	return key, err
-}
